Feature binary (#206)
* Support Historical Source Recovery
----------------------------------
This applies to an individual element and is not a git recovery but
a source level recovery. This only works for z/OS sequential datasets
and PDS members (one at a time). OMVS files are not support at this
time but may be in the future.

* Update Rollback  dialog - new function
--------------------------------------
New commands for rollback:
- Only will limit the table display to those commits that match the
provided string (only string) in the tag or title
- Refresh will recreate the table

This for repositories with a large number of commits to make it easier
to find a specific commit.

* Clarify the Rollback update
---------------------------
1. only look for the only string in the title
2. update the release notes for only/refresh in rollback

* cosmetic update to zigi command panel

* Two updates
-----------
1. Update rollback panel - cosmetic
2. Update selective history recovery
- if updating the active dataset call update_repo_metadata
- fix bug in the panel rexx code

* History Recovery bug fix
------------------------
more bugs in the recovery code in the panel rexx code fixed

* One more update for history recover
-----------------------------------
After the recovery to the active pds the update_repo_metadata was not
the right routine - changed to work_with_repo_file

* one last (hopefully) fix for historical recover

* Correct bug in adding files with special chars

* Small updates
-------------
1. shift the recovery prompt popup over to column 8 from 12
2. update the history / (row selection) popup for Recover

Co-authored-by: Lionel B. Dyck <42328411+lbdyck@users.noreply.github.com>