Multiple updates ---------------- Allow a qualifier ignore value of zero to allow datasets from different HLQ's to be in the same repository. Not a great idea but doable. The ful dataset name (e.g. SYS1.PARMLIB) will now be stored in the repository. When doing this care is required to prevent updating the z/OS dataset fr the OMVS filesystem. When cloning such a repository an HLQ would still b required (thus SYS1.PARMLIB would clone into hlq.SYS1.PARMLIB).
When changing branches, or pulling an update, if a OMVS file was
removed then the z/OS dataset that maps to it is not removed. If the
z/OS dataset was a PDS then the members were deleted. The member
deletes will no longer happen.

Updated the add dataset popup panel to add a command entry field. This
allows a file extension (which was the 1st entry field) to have values
that may map to a ISPF command table entry.

After a Merge the current repository display needed to be updated to
Generalize binary file detection in case someone modifies .gitattributes
manually and doesn't follow the ZIGI convention.