Let's go, three dot o two (#380)
* Add _EDC_ZERO_RECLEN=Y better VB support
----------------------------------------
Add _EDC_ZERO_RECLEN=Y to the OMVS environment variables to address
datasets/members with variable length records (RECFM=VB) with blank
records.  This will preserve those records.

* Remove redundant repo delete code
---------------------------------
Duplicate code resulted in erroneous message that the OMVS files
had not been deleted when they had.

* Update Help->About with youtube link

* Add calls to enable/disable pfkey displays
------------------------------------------
Add to all ADDPOPs that were missing this

* Multiple updates
----------------
Allow a qualifier ignore value of zero to allow datasets from different
HLQ's to be in the same repository. Not a great idea but doable. The ful
dataset name (e.g. SYS1.PARMLIB) will now be stored in the repository.
When doing this care is required to prevent updating the z/OS dataset fr
the OMVS filesystem. When cloning such a repository an HLQ would still b
required (thus SYS1.PARMLIB would clone into hlq.SYS1.PARMLIB).

When changing branches, or pulling an update, if a OMVS file was
removed then the z/OS dataset that maps to it is not removed. If the
z/OS dataset was a PDS then the members were deleted. The member
deletes will no longer happen.

Updated the add dataset popup panel to add a command entry field. This
allows a file extension (which was the 1st entry field) to have values
that may map to a ISPF command table entry.

After a Merge the current repository display needed to be updated to
Generalize binary file detection in case someone modifies .gitattributes
manually and doesn't follow the ZIGI convention.

* Initial 3.02 Development Drop
-----------------------------
Support for a Read Only Repository.

* Add check for z/OS dataset replacement and prompt

* Prevent ignr change from 0

* Update on replace notification to bypass prompt
-----------------------------------------------
bypass prompt if doing selective replace.

**** THIS NEEDS WORK ****

* After Stash Pop leave stash routine
-----------------------------------
After the pop there is no need to remain in the stashlist table display

* Only Prompt for z/OS replace on Clone or Replace

* Revise Replace Prompt to Clone only

* Correct wording in ZIGICDM panel

* Clone of non-ZIGI Repo bypass gitattributes +
---------------------------------------------
and do not create .zigi/dsn

* Correct ReadOnly and Member Delete handling

* Fix file tag got .gitignore

* Multiple Updates for user friendly experience
---------------------------------------------
1. if qualignr=0 then don't save the prefix (hlq)
2. if qualignr=0 correctly detect the added datasets in the add table
3. clarify and simplify the delete repo from repository panel

* Support Multiple Prefixes (HLQs) on dsn add

* Update del repo processing
--------------------------
When the panel questions were changed Y used to mean delete
now it means KEEP

* One more fix to get Delete Repo right

* Fix ZIGIREFD temp ds alloc for No TSO Prefix

* Add tutorial on readonly repos

* Update stat compare for the future

* Update stat checking one more time

* Correct typo in release notes

* Performance update after updating PDS member

Co-authored-by: Lionel.Dyck <lioneld@21csw.com>