V3r10 (#399)
* Add _EDC_ZERO_RECLEN=Y better VB support
----------------------------------------
Add _EDC_ZERO_RECLEN=Y to the OMVS environment variables to address
datasets/members with variable length records (RECFM=VB) with blank
records.  This will preserve those records.

* Remove redundant repo delete code
---------------------------------
Duplicate code resulted in erroneous message that the OMVS files
had not been deleted when they had.

* Update Help->About with youtube link

* Add calls to enable/disable pfkey displays
------------------------------------------
Add to all ADDPOPs that were missing this

* Multiple updates
----------------
Allow a qualifier ignore value of zero to allow datasets from different
HLQ's to be in the same repository. Not a great idea but doable. The ful
dataset name (e.g. SYS1.PARMLIB) will now be stored in the repository.
When doing this care is required to prevent updating the z/OS dataset fr
the OMVS filesystem. When cloning such a repository an HLQ would still b
required (thus SYS1.PARMLIB would clone into hlq.SYS1.PARMLIB).

When changing branches, or pulling an update, if a OMVS file was
removed then the z/OS dataset that maps to it is not removed. If the
z/OS dataset was a PDS then the members were deleted. The member
deletes will no longer happen.

Updated the add dataset popup panel to add a command entry field. This
allows a file extension (which was the 1st entry field) to have values
that may map to a ISPF command table entry.

After a Merge the current repository display needed to be updated to
Generalize binary file detection in case someone modifies .gitattributes
manually and doesn't follow the ZIGI convention.

* Initial 3.02 Development Drop
-----------------------------
Support for a Read Only Repository.

* Add check for z/OS dataset replacement and prompt

* Prevent ignr change from 0

* Update on replace notification to bypass prompt
-----------------------------------------------
bypass prompt if doing selective replace.

**** THIS NEEDS WORK ****

* After Stash Pop leave stash routine
-----------------------------------
After the pop there is no need to remain in the stashlist table display

* Only Prompt for z/OS replace on Clone or Replace

* Revise Replace Prompt to Clone only

* Correct wording in ZIGICDM panel

* Clone of non-ZIGI Repo bypass gitattributes +
---------------------------------------------
and do not create .zigi/dsn

* Correct ReadOnly and Member Delete handling

* Fix file tag got .gitignore

* Multiple Updates for user friendly experience
---------------------------------------------
1. if qualignr=0 then don't save the prefix (hlq)
2. if qualignr=0 correctly detect the added datasets in the add table
3. clarify and simplify the delete repo from repository panel

* Support Multiple Prefixes (HLQs) on dsn add

* Update del repo processing
--------------------------
When the panel questions were changed Y used to mean delete
now it means KEEP

* One more fix to get Delete Repo right

* Fix ZIGIREFD temp ds alloc for No TSO Prefix

* Add tutorial on readonly repos

* Update stat compare for the future

* Update stat checking one more time

* Correct typo in release notes

* Performance update after updating PDS member

* V3R10 updates
-------------
Pulled then verified and updated for 3.10 release - now release candidat

* 393 bug fix - repo delete missing z/os datasets

* Issue 395: VB datasets in Binary are corrupted
----------------------------------------------
The OMVS cp command does NOT support variable length binary datasets
causing the corruption. Updating ZIGI to prevent adding them to the
git repository.

* Improve recfm=v binary error message w/dsn

* Add Copy popup status when copying PS datasets
----------------------------------------------
Just to improve awareness when doing this copy

* Correct incorrect zigidebug initialization

* Several minor updates
---------------------
If a PDS member is missing but it's counterpart if found in the
   OMVS filesystem then:
1. Correct the table panel display
2. Correct the rm's

* Correct table handling after CHECK proessing

* Update ISPF stats for all members

* Update ADDALL to  git add .  - much faster

* Add command to message view

* Fix dialog failure on scroll max on history view

* Fixed ZIGIEDIT panel for max scrolling
--------------------------------------
Fixed the panel rexx code for the length of zshadow and restored
the use of the zigiedit panel for history viewing

* Update ispf stats

* Fix bug with lost stats after a pull
------------------------------------
Problem in ZIGICKOT when a pull had only a few members changed
the code wasn't properly handling the Dovetail putpds in this
case but was ignoring the ispf stat update as the putpds will
do that.  Now fixed.

* Fix TBSORT on the repo table if fields var bad

* Fix return from omvs subdirectories
-----------------------------------
A long lasting bug has now been fixed. It would take 2+ F3's to
return from an omvs subdirectory. Now it works the way it should have.

Co-authored-by: Lionel.Dyck <lioneld@21csw.com>
Co-authored-by: Lionel.Dyck <lbdyck@gmail.com>