3.24 (#564)
* Check for working with current EXEC
-----------------------------------
Check the name of the EXEC used to invoke ZIGI. If you're trying to upda
the repository that uses this EXEC, we stop immediately and tell you thi
is a bad idea. You really should use a "prod" ZIGI to work on a dev ZIGI

* correct zginstall zgstat for 8 char userids

* ZIGI Generalized Installation Tool added.

* update zginstal to honor nocoz dd

* Update zginstal
---------------
failed to test for NOCOZ as a test option to force
using omvs cp instead of co:z

* change to v3.24 in zigi version
-------------------------------
version change only (so far

* updates to set 3.24 as new version

* really set 3.24 as the starting current ver

* Correct detection of ssh key or lack of

* Update to improve the Branch display info
-----------------------------------------
add userid and date of last update for branch displays
- BRanch command
- MERge command
- DB (diff branch selection)

* update zigi
-----------
add member stat variable initialization to null

* Correct usage of ssh-keygen
---------------------------
Use the z/OS supplied one in /bin
- to prevent using one that incorrectly tags the keys
  as git/GitHub needs the keys to be untagged

* Sync for end of year

* Update self-extracting panels

* Panels updated

* missed panels

* Add update data to ZG

* fix updated panel userids

* update ispf stats for exec

* Update for ssh key generation

* update window size on two popups

* update more popup panels
------------------------
adjust window size and/or )panel window syntax

* improved rsa key support and ed25519 keys

* update ssh key type/size prompt

* fix ssh key prompt from ed12259 to ed25519

* update ssh-keygen
-----------------
check for which version of openssh (ssh-keygen) is installed
and if the z/OS version then only support RSA keys at 3072 bits

* improve prompting for ssh key
-----------------------------
- if not zopen openssh then only prompt for rsa
- if zopen openssh then prompt for rsa and ed25519
  with ed25519 being the default

* close but . . .
---------------
issues with the ed25519 keys as it reportws id_ed25519 not found
BUT it works great for RSA keys

* ssh regen working
-----------------
well mostly - always for rsa but on ed25519 not so much
if switching from ed to rsa during initial setup

* clean up ssh regen issue
------------------------
during initial setup a regen will fail if changing from
rsa to ed25519 or vice versa - now fixed

* Update release info regarding ssh keys
--------------------------------------
ssh keys generated with type of ed25519 prior to the installation
of the zopen openssh are untagged and will fail after the zopen
openssh is installed - the solution is to recreate the keys.
IF the zopen ssh-keygen is not installed then only RSA ssh keys
are supported.

* update test of ssh-keygen version
---------------------------------
use ssh -V to get the version and if 9 or greater then
support ssh key of ed25519, otherwise only support rsa keys

* update release doc

* update tutorial for openssh version

* correct initial ssh version detection

* Change z/OS Open Tools to zopen community

* refine ssh key processing

* update ssh key detection
------------------------
if both rsa and ed25519 keys found then ask which to use

* correct ssh key detection and prompt

* refine ssh key selection

* one more ssh key detection refinement

* update release notes with todays date

* small updates
-------------
- display ssh key filename in GitInfo

* add patience popup for ssh-keygen
---------------------------------
just to let the user know it is running
- will be fairly quick for ed25519 and rsa 2048
  but longer for rsa 3072 and 4096

* two bug fixes
-------------
set GITENV to null
fix zigirset panel

* bug fixes
---------
set GITENV variable in zigi exec to null to start with
fix bug in zigirset panel

* more fix to mixed pds with cp bug

* Update zigistat
---------------
to correct a bug when Co:Z is not installed where zigistat
attempted to read ISPF statistics rather than performing
a ispf statistics update on a newly, cp created, pds.

* Update zigickot
---------------
improved the copy (putpds) by eliminating poorly architected
copy

* update zigickot
---------------
fixed bug with nocoz on mixed pds
.
BUT still needs work - the sed to cleanup routine needs to
be for both nocoz and coz
AND the cp for text and bin should be enhanced to match
the mixed coz with one command for text and one for bin
.
It works for now.

* finished update to zigickot
---------------------------
cleaned up mixed pds processing for both cp and co:z

* Testing with updated zigiedit panel
-----------------------------------
changed ispf panel attributes 05 to 35 and 0D to 3D

* update release notes

* fix diff all report

* all branch displays do not include current

* Fix bug 559
-----------
Command BR conflicts with user command table - fixing in pulldown
menu from BR to BRANCH. The code supports any validate abbreviation
of BRANCH from BR to BRANCH.

* 3.24 update
-----------
Add the latest version/release info to the Help/About panel

* Update for release date of Mar 14 2025

---------

Co-authored-by: Ray Mullins <m.ray.mullins@gmail.com>
Co-authored-by: Lionel B. Dyck <42328411+lbdyck@users.noreply.github.com>
Co-authored-by: Lionel.Dyck <lbdyck@gmail.com>
Co-authored-by: Igor Todorovski <39890068+IgorTodorovskiIBM@users.noreply.github.com>