3.25 (#581)
* v3r25 initial commit
--------------------
- fix completion message on ssh key regen

* Correct non-Co:Z handling of pds aliases
----------------------------------------
PDS member aliases are ignored by Co:Z and should
be ignored by non-Co:Z uses. The real member is processed
and there is no way to identify the alias to git.

* v3.25 update for zg
-------------------
improve the test for getpds if co:z not available

* Create known_hosts if it doesn't exist
--------------------------------------
If the ~/.ssh/known_hosts does not exist then create it with a copy
of the local public key using github.com as the initial entry. This
should be fine for other servers as well as it just needs to exist.

* Improve update of known_hosts
-----------------------------
use ssh -o StrictHostKeyChecking=no git@??????.??? -q
where the ??????.??? is the host server

* don t try to update known_hosts if no remote

* change ssh key to accept-new
----------------------------
Thanks Igor

* update panel zigisshk to allow wide screen

* Clean up and improve known_hosts update

* remove extraneous line of code
------------------------------
sigh

* Update to remove openssh v9 check

* clean up known_hosts update
---------------------------
only do if the host is not in the file

* cosmetic change to zigisshk panel

* remove HEAD from branch lists
-----------------------------
redundant so not needed

---------

Co-authored-by: lionel.dyck <lbdyck@gmail.com>
Co-authored-by: Lionel B. Dyck <42328411+lbdyck@users.noreply.github.com>